requirements.txt
hinlangpy>=1.0.0
PyQt6>=6.6.0
pyinstaller>=6.0.0
